requirements.txt
streamlit
python-dotenv
Pillow
google-generativeai
streamlit-option-menu
pandas
numpy